move sketch new test in separate test_new.py file
